Fix application context so that Spring and Cucumber play nicely with multiple step definition files.
